cassandra done

+++ sql/postGIStables.sql
@@ -2,7 +2,7 @@ CREATE TABLE unoptimized(
     id SERIAL PRIMARY KEY,
     severity INT,
     us_state VARCHAR(2),
-    precipitation INT, 
+    precipitation FLOAT, 
     windy SMALLINT, 
     geo POINT,
     start_time DATE, 
@@ -14,7 +14,7 @@ CREATE TABLE optimized(
     severity INT,
     us_state VARCHAR(2),
     windy SMALLINT
-    precipitation INT, 
+    precipitation FLOAT, 
     geo POINT,
     start_time DATE, 
     end_time DATE
